Practical Machine Learning Dissertation
Practical Machine Learning Dissertation  assignment first submission. Not sure if R code is allowed to be shared or the assignment simply needs the (intermidiate/final) results of the model, theoretical justification and explanation of the thought process.
Once confirmed with the  course Tutors I can submit the R code or  structure it in a way that only include the R model outputs.